+++ language/en/administrator.php
@@ -48,4 +48,5 @@ return array(
 	'valid_config_path' => "You must provide a valid 'model_config_path' in the administrator.php config. The directory must also exist and be readable.",
 	'not_eloquent' => " is not an Eloquent model",
 	'storage_path_permissions' => "You must make your storage path writable in order to make a settings page",
+	'valid_home_page' => "You must provide a valid menu item in the home_page option of your config",
 );